# api/product_discount.go
package handler

import (
	"database/sql"
	"encoding/json"
	"fmt"
	"net/http"
	"price-control/api/generated"
	"price-control/pkg/db"
)

var dbConnection, _ = db.ConnectToDb()

func ProductDiscountHandler(w http.ResponseWriter, request *http.Request) {
	switch request.Method {
	case http.MethodPost:
		SetProductDiscount(w, request)
	case http.MethodGet:
		GetDiscountedProducts(w, request)
	}
}

func GetDiscountedProducts(w http.ResponseWriter, r *http.Request) {
	discountType := r.URL.Query().Get("discountType")
	if discountType == "" {
		http.Error(w, "Missing discountType parameter", http.StatusBadRequest)
		return
	}

	query := "SELECT name FROM products WHERE discount = $1"
	rows, err := dbConnection.Query(query, discountType)
	if err != nil {
		http.Error(w, "Error querying database: "+err.Error(), http.StatusInternalServerError)
		return
	}
	defer func(rows *sql.Rows) {
		_ = rows.Close()
	}(rows)

	var products []string
	for rows.Next() {
		var productName string
		if err := rows.Scan(&productName); err != nil {
			http.Error(w, "Error scanning row: "+err.Error(), http.StatusInternalServerError)
			return
		}
		products = append(products, productName)
	}

	if err := rows.Err(); err != nil {
		http.Error(w, "Error with rows: "+err.Error(), http.StatusInternalServerError)
		return
	}

	response := map[string][]string{"products": products}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	_ = json.NewEncoder(w).Encode(response)
}

func SetProductDiscount(w http.ResponseWriter, request *http.Request) {
	var product generated.PostProductDiscountJSONBody

	err := json.NewDecoder(request.Body).Decode(&product)
	if err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	if product.DiscountType == "" || product.ProductName == "" {
		http.Error(w, "Discount type and product name cannot be empty", http.StatusBadRequest)
		return
	}

	if err != nil {
		_, _ = fmt.Fprintf(w, "<h1>Error connecting to database: %v(</h1>", err)
		return
	}

	defer func(dbConnection *sql.DB) {
		_ = dbConnection.Close()
	}(dbConnection)

	query := "UPDATE products SET discount = $1 WHERE name = $2"
	fmt.Println(query)
	_, err = dbConnection.Exec(query, product.DiscountType, product.ProductName)
	if err != nil {
		_, _ = fmt.Fprintf(w, "<h1>Error inserting product discount: </h1>")
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusOK)
	w.Header().Set("Content-Type", "application/json")
	_, _ = w.Write([]byte(`{"message": "Product discount added successfully"}`))
}

func init() {
	http.HandleFunc("/productDiscount", ProductDiscountHandler)
}